Change to Refcounting semantics
In the previous commit, once you called AsObservable(), we kept trucking on
forever even if the last subscriber dropped off early, until the stream was
exhausted.

Instead, have at most one underlying person doing the reads and the rest
recieving that result, and if everybody unsubscribes early, we don't Dispose
and we stop reading the stream altogether.